requirements.txt
flask
python-dotenv
parallel-web
httpx
gunicorn
supabase
streamlit
openai
google-api-python-client
yt-dlp
pydub
sarvamai
isodate
